MongoDB integration, verification re-write, more embeds

+++ src/main/java/io/paradaux/csbot/listeners/message/VerificationEmailReceivedListener.java
@@ -24,9 +24,8 @@
 package io.paradaux.csbot.listeners.message;
 
 import io.paradaux.csbot.api.ConfigurationCache;
-import io.paradaux.csbot.api.SMTPConnection;
-import io.paradaux.csbot.api.VerificationSystem;
 import io.paradaux.csbot.controllers.ConfigurationController;
+import io.paradaux.csbot.controllers.DatabaseController;
 import io.paradaux.csbot.controllers.EmailController;
 import io.paradaux.csbot.controllers.LogController;
 import net.dv8tion.jda.api.entities.Message;
@@ -39,52 +38,57 @@ import javax.mail.MessagingException;
 
 public class VerificationEmailReceivedListener extends ListenerAdapter {
 
-    ConfigurationCache configurationCache;
-    SMTPConnection smtpConnection;
-    Logger logger;
-
-    public VerificationEmailReceivedListener() {
-        this.configurationCache = ConfigurationController.getConfigurationCache();
-        this.smtpConnection = EmailController.getSmtpConnection();
-        logger = LogController.getLogger();
-    }
+    ConfigurationCache configurationCache = ConfigurationController.getConfigurationCache();
+    DatabaseController databaseController = DatabaseController.INSTANCE;
+    Logger logger = LogController.getLogger();
 
     @Override
     public void onMessageReceived(@NotNull MessageReceivedEvent event) {
 
-        if (!event.getChannel().getId().equals(configurationCache.getListeningChannel())) return;
-
         Message message = event.getMessage();
-        message.delete().queue();
+        String discordID = event.getAuthor().getId(), guildID = event.getGuild().getId();
         String email = message.getContentRaw();
+
+        if (event.getAuthor().isBot()) return;
+        if (!event.getChannel().getId().equals(configurationCache.getListeningChannel())) return;
+        if (databaseController.isPendingVerification(discordID)) return;
+
+        // We need to handle the message deletion.
+        message.delete().queue();
 
         if (!EmailController.isValidEmail(email)) {
             event.getAuthor().openPrivateChannel().queue((channel) -> {
-                channel.sendMessage("Your message did not contain an email.").queue();
+                channel.sendMessage("Your message did contain a valid email address.").queue();
             });
-
             return;
         }
 
-        if (!EmailController.getEmailDomain(email).equalsIgnoreCase("tcd.ie")) {
+        // If it isn't an @tcd.ie email
+        if (!EmailController.getEmailDomain(email).equals("tcd.ie")) {
             event.getAuthor().openPrivateChannel().queue((channel) -> {
-                channel.sendMessage("The provided email must be of the tcd.ie domain. If you are not a trinity student, please contact a moderator for manual verification").queue();
+                channel.sendMessage("You must use a valid @tcd.ie email address to go through automatic verification." +
+                        "\nIf you are not a trinity student please respond to the bot in this channel and a moderator will be with you shortly." +
+                        "\nPlease message the bot if you run into issues with this, a moderator/technician will be with you shortly.").queue();
             });
 
-            return;
-        };
+        }
 
-        event.getAuthor().openPrivateChannel().queue((channel) -> {
-            channel.sendMessage("Please check your email for a verification token. Once you have it, please paste it into this private message channel.").queue();
-        });
-
-        VerificationSystem.addPendingUser(event.getAuthor().getId(), EmailController.generateVerificationCode());
+        // Send a verification code to the user
+        String verificationCode = EmailController.generateVerificationCode();
         try {
-            smtpConnection.sendVerificationEmail(email, VerificationSystem.getVerificationCode(event.getAuthor().getId()));
+            EmailController.INSTANCE.sendVerificationEmail(email, verificationCode, event.getAuthor().getAsTag());
         } catch (MessagingException exception) {
-            logger.error("There was an error sending the verification email for {}", event.getAuthor().getName(), exception);
+            logger.error("Error sending email to {} with email address {}", event.getAuthor().getAsTag(), email, exception);
             return;
         }
 
+        // Add the user pending verification to the database.
+        databaseController.addPendingVerificationUser(discordID, guildID, verificationCode);
+
+        // Notify the user that there's an email waiting for them.
+        event.getAuthor().openPrivateChannel().queue((channel) -> {
+            channel.sendMessage("Please check your email for a verification token. Once you have it, please paste it back into #verification." +
+                    "\nPlease message the bot if you run into issues with this, a moderator/technician will be with you shortly.").queue();
+        });
     }
 }
